requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
boto3==1.14.49
    # via -r requirements.in
botocore
    # via
    #   boto3
    #   s3transfer
click==8.0.3
    # via pip-tools
cssselect2==0.4.1
    # via svglib
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.13
    # via
    #   -r requirements.in
    #   django-classy-tags
    #   django-cms
    #   django-formtools
    #   django-sekizai
    #   django-storages
    #   django-treebeard
    #   easy-thumbnails
django-classy-tags==2.0.0
    # via
    #   -r requirements.in
    #   django-cms
    #   django-sekizai
django-cms==3.10.0
    # via
    #   -r requirements.in
    #   djangocms-attributes-field
    #   djangocms-bootstrap4
    #   djangocms-file
    #   djangocms-googlemap
    #   djangocms-icon
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-style
    #   djangocms-text-ckeditor
    #   djangocms-video
django-crispy-forms==1.14.0
    #   added manually on 2023-02-08
    #   required for password resets
django-filer @ https://github.com/django-cms/django-filer/archive/refs/tags/2.1.1.zip
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
    #   djangocms-file
    #   djangocms-googlemap
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-video
django-formtools==2.3
    # via django-cms
git+http://git@github.com/divio/django-polls.git#egg=polls
django-utils-six==2.0
    #   added manually on 2023-02-08
    #   required by token authentication (deprecated)
django-sekizai==2.0.0
    # via
    #   -r requirements.in
    #   django-cms
django-storage-url==0.5.0
    # via -r requirements.in
django-storages==1.11.1
    # via django-storage-url
django-treebeard==4.5.1
    # via django-cms
djangocms-admin-style==3.1.1
    # via
    #   -r requirements.in
    #   django-cms
djangocms-attributes-field==2.0.0
    # via
    #   djangocms-bootstrap4
    #   djangocms-file
    #   djangocms-icon
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-style
    #   djangocms-video
djangocms-bootstrap4==2.0.0
    # via -r requirements.in
djangocms-file==3.0.0
    # via -r requirements.in
djangocms-googlemap==2.0.0
    # via -r requirements.in
djangocms-icon==2.0.0
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-link==3.0.0
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-picture==3.0.0
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-style==3.0.0
    # via -r requirements.in
djangocms-text-ckeditor==4.0.0
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-video==3.0.0
    # via -r requirements.in
django-dbbackup ==4.0.2
docutils==0.15.2
    # via botocore
easy-thumbnails==2.8.0
    # via
    #   -r requirements.in
    #   djangocms-picture
furl==2.1.3
    # via django-storage-url
html5lib==1.1
    # via djangocms-text-ckeditor
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
lxml==4.6.3
    # via svglib
orderedmultidict==1.0.1
    # via furl
pep517==0.12.0
    # via pip-tools
pillow==8.3.2
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pip-tools==6.4.0
    # via -r requirements.in
psycopg2==2.8.5
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
pytz==2021.1
    # via
    #   -r requirements.in
    #   django
reportlab==3.6.1
    # via
    #   easy-thumbnails
    #   svglib
s3transfer==0.3.7
    # via boto3
six==1.16.0
    # via
    #   -r requirements.in
    #   furl
    #   html5lib
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.4.2
    # via django
svglib==1.1.0
    # via easy-thumbnails
tinycss2==1.1.0
    # via
    #   cssselect2
    #   svglib
tomli==1.2.2
    # via pep517
urllib3
    # via botocore
uwsgi==2.0.19.1
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.37.0
    # via pip-tools
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
